auth(4): database seeder with admin account, fix hydration issues in navbar after login/logout

+++ server/api/session.delete.ts
@@ -1,22 +1,18 @@
 import { sessions } from '~~/db/schema';
 import { db } from './utils/database';
-import { validateSessionToken } from './utils/session';
 import { eq } from 'drizzle-orm';
 
 export default defineEventHandler(async (event) => {
   try {
-    const sessionToken = getCookie(event, 'session_token');
-    if (sessionToken) {
-      const validSessionToken = await validateSessionToken(db, sessionToken);
-      if (validSessionToken) {
-        await db.delete(sessions).where(eq(sessions.id, validSessionToken.id));
-        deleteCookie(event, 'session_token');
+    if (event.context.auth.validSessionToken) {
+      await db.delete(sessions).where(eq(sessions.id, event.context.auth.validSessionToken.id));
+      deleteCookie(event, 'session_token');
+      deleteCookie(event, 'user');
 
-        return {
-          success: true,
-          error: null,
-        };
-      }
+      return {
+        success: true,
+        error: null,
+      };
     }
     return {
       success: false,
